+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to Rich Foot will be documented in this file.
 
+## [1.9.2] - 2024-12-01
+### 🐛 Fixed
+- dynamic `css` in `reading` mode disrupting document flow of floated elements (e.g. ITS callouts)
+- debounced `updateRichFoot` in `editing` mode (new settings option that allows delay in milliseconds)
+
 ## [1.9.1] - 2024-12-01
 ### 🐛 Fixed
 - `Links` defined in frontmatter were not being displayed